restructure codes to posts
