+++ src/js/api/posts/index.mjs
@@ -1,4 +1,4 @@
 export * from './create.mjs';
 export * from './read.mjs';
 export * from './update.mjs';
-export * from './delete.mjs';
+export * from './remove.mjs';
